QuantityTest description
